refactor: reorganize static assets and modules
