+++ assets/styles/components/hw-panel.css
