+++ index.html
@@ -4,12 +4,12 @@
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1, viewport-fit=cover" />
   <title>作业预览 · 苹果拟物</title>
-  <link rel="stylesheet" href="/styles/tokens.css" />
-  <link rel="stylesheet" href="/styles/ios-theme.css" />
-  <link rel="stylesheet" href="/css/base.css" />
-  <link rel="stylesheet" href="/css/layout.css" />
-  <link rel="stylesheet" href="/css/hw-panel.css" />
-  <link rel="stylesheet" href="/css/donut.css" />
+  <link rel="stylesheet" href="./assets/styles/theme/tokens.css" />
+  <link rel="stylesheet" href="./assets/styles/theme/ios-theme.css" />
+  <link rel="stylesheet" href="./assets/styles/base/base.css" />
+  <link rel="stylesheet" href="./assets/styles/base/layout.css" />
+  <link rel="stylesheet" href="./assets/styles/components/hw-panel.css" />
+  <link rel="stylesheet" href="./assets/styles/components/donut.css" />
 </head>
 <body>
   <div class="app" id="app">
@@ -34,7 +34,6 @@
       </div>
     </div>
   </div>
-  <script defer src="scripts/flip.js"></script>
-  <script type="module" src="/src/app.js"></script>
+  <script type="module" src="./src/main.js"></script>
 </body>
 </html>
