+++ src/main.js
@@ -1,0 +1,1 @@
+import './core/app.js';
