+++ src/modules/homework/index.js
@@ -1,14 +1,22 @@
 import { render, updateCompletion } from './render.js';
-import { debounce } from '../utils/dom.js';
-import { flipReorder, animateAutoHeight } from '/scripts/flip.js';
-import { state, getState, setState, addSubject as add, removeSubject as remove, updateTask } from './state.js';
+import { debounce } from '../../shared/dom.js';
+import { flipReorder, animateAutoHeight } from '../../animations/flip.js';
+import {
+  state,
+  getState,
+  setState,
+  addSubject as add,
+  removeSubject as remove,
+  updateTask,
+  selectProgress
+} from './state.js';
 import { sortByCompleteThenSeq } from './sort.js';
-import { selectProgress } from './state.js';
 
 export function initHwPanel({ mount, onProgress }) {
   render(mount, state);
   updateCompletion(mount, state);
   onProgress(selectProgress());
+
   const debounced = debounce(() => {
     animateAutoHeight(mount, 180);
     flipReorder(mount, '.card', () => {
@@ -16,6 +24,7 @@ export function initHwPanel({ mount, onProgress }) {
       items.forEach(el => mount.appendChild(el));
     }, { duration: 300, easing: 'cubic-bezier(.2,.8,.2,1)', stagger: 24 });
   }, 100);
+
   mount.addEventListener('change', e => {
     const input = e.target;
     if (input && input.matches('input[type="checkbox"]')) {
@@ -28,6 +37,7 @@ export function initHwPanel({ mount, onProgress }) {
       onProgress(selectProgress());
     }
   });
+
   return { addSubject, removeSubject, updateTask, getState, setState };
 }
